+++ test-suite/TestExitCode/Program.cs
@@ -64,10 +64,8 @@ class Program
         }
         else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
         {
-            // Note: diagnostics fail to retrieve the exit code on macOS; skipping this test for now.
-
-            //Console.WriteLine("Test _exit()");
-            //_exit(3);
+            Console.WriteLine("Test _exit()");
+            _exit(3);
         }
         // Console.WriteLine("Test return 3");
         // return 3;
@@ -79,7 +77,7 @@ class Program
             (Object context) =>
             {
                 Context Context = (Context)context;
-                Context.WasExit(RuntimeInformation.IsOSPlatform(OSPlatform.OSX) ? 0 : 3, @"__FILE__:__LINE__");
+                Context.WasExit(3, @"__FILE__:__LINE__");
                 Context.DebuggerExit(@"__FILE__:__LINE__");
             });
     }