+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Upcoming changes compared to previous version.
 - Fixed Windows x86 build.
 - Fixed macOS arm64 and x86_64 builds (iconv proper detection).
 - Fixed Windows arm64 architecture detection during the build process.
+- Fixed macOS child process exit code retrieval by using kqueue with NOTE_EXITSTATUS.
 
 </br>
 </br>